refac(ft_cd.c): just separate file ft_cd_update_pwd
